Crate ink_primitives

Utilities in use by ink!.

These are kept separate from ink! core utilities to allow for more dynamic inter-crate dependencies. The main problem is that today Cargo manages crate features on a per-crate basis instead of a per-crate-target basis thus making dependencies from ink (or others) to ink_env or ink_storage impossible.

By introducing ink_primitives we have a way to share utility components between ink_env or ink_storage and other parts of the framework, like ink.

Re-exports

pub use self::types::AccountId;

pub use self::types::Clear;

pub use self::types::DepositLimit;

pub use self::types::Hash;

Modules

contract

reflect

Definitions and utilities for ink! smart contract static reflection.

types

Structs

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

KeyComposer

Contains all rules related to storage key creation.

U256

Little-endian large integer type 256-bit unsigned integer.

Enums

LangError

An error emitted by the smart contracting language.

Type Aliases

ConstructorResult

The Result type for ink! constructors.

Key

A key into the smart contract storage.

MessageResult

The Result type for ink! messages.