ink_primitives
Module contract

Traits

ContractEnv

Stores the used host environment type of the ink! smart contract.

ContractReference

Refers to the generated ink! smart contract reference type.

ContractReverseReference

Refers back to the original contract from the generated ink! smart contract reference type.