ink_primitives::sol

Function decode_sequence

pub fn decode_sequence<T: SolParamsDecode>(data: &[u8]) -> Result<T, Error>

Solidity ABI decode the given data as a parameter sequence.

Note

T must be a tuple type where each member implements SolDecode.

See notes for encode_sequence for the difference between this function and SolDecode::decode for the given tuple.