ink_primitives
Struct U256

#[repr(C)]
pub struct U256(pub [u64; 4]);

Little-endian large integer type 256-bit unsigned integer.

Tuple Fields
0: [u64; 4]
Implementations

impl U256

pub const MAX: U256 = _

Maximum value.

pub fn from_str_radix(txt: &str, radix: u32) -> Result<U256, FromStrRadixErr>

Converts a string slice in a given base to an integer. Only supports radixes of 10 and 16.

pub fn from_dec_str(value: &str) -> Result<U256, FromDecStrErr>

Convert from a decimal string.

pub const fn low_u32(&self) -> u32

Conversion to u32

pub const fn low_u64(&self) -> u64

Low word (u64)

pub fn as_u32(&self) -> u32

Conversion to u32 with overflow checking

Panics

Panics if the number is larger than 2^32.

pub fn as_u64(&self) -> u64

Conversion to u64 with overflow checking

Panics

Panics if the number is larger than u64::max_value().

pub fn as_usize(&self) -> usize

Conversion to usize with overflow checking

Panics

Panics if the number is larger than usize::max_value().

pub const fn is_zero(&self) -> bool

Whether this is zero.

pub fn bits(&self) -> usize

Return the least number of bits needed to represent the number

pub const fn bit(&self, index: usize) -> bool

Return if specific bit is set.

Panics

Panics if index exceeds the bit width of the number.

pub fn leading_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_zeros(&self) -> u32

Returns the number of trailing zeros in the binary representation of self.

pub const fn byte(&self, index: usize) -> u8

Return specific byte. Byte 0 is the least significant value (ie~ little endian).

Panics

Panics if index exceeds the byte width of the number.

pub fn to_big_endian(&self) -> [u8; 32]

Convert to big-endian bytes.

pub fn write_as_big_endian(&self, bytes: &mut [u8])

Write to the slice in big-endian format.

pub fn to_little_endian(&self) -> [u8; 32]

Convert to little-endian bytes.

pub fn write_as_little_endian(&self, bytes: &mut [u8])

pub fn exp10(n: usize) -> U256

Create 10**n as this type.

Panics

Panics if the result overflows the type.

pub const fn zero() -> U256

Zero (additive identity) of this type.

pub const fn one() -> U256

One (multiplicative identity) of this type.

pub const fn max_value() -> U256

The maximum value which can be inhabited by this type.

pub fn div_mod(self, other: U256) -> (U256, U256)

Returns a pair (self / other, self % other).

Panics

Panics if other is zero.

pub fn integer_sqrt(&self) -> U256

Compute the highest n such that n * n <= self.

pub fn pow(self, expon: U256) -> U256

Fast exponentiation by squaring https://en.wikipedia.org/wiki/Exponentiation_by_squaring

Panics

Panics if the result overflows the type.

pub fn overflowing_pow(self, expon: U256) -> (U256, bool)

Fast exponentiation by squaring. Returns result and overflow flag.

pub fn checked_pow(self, expon: U256) -> Option<U256>

Checked exponentiation. Returns None if overflow occurred.

pub fn overflowing_add(self, other: U256) -> (U256, bool)

Addition which overflows and returns a flag if it does.

pub fn saturating_add(self, other: U256) -> U256

Addition which saturates at the maximum value (Self::MAX).

pub fn checked_add(self, other: U256) -> Option<U256>

Checked addition. Returns None if overflow occurred.

pub fn overflowing_sub(self, other: U256) -> (U256, bool)

Subtraction which underflows and returns a flag if it does.

pub fn saturating_sub(self, other: U256) -> U256

Subtraction which saturates at zero.

pub fn checked_sub(self, other: U256) -> Option<U256>

Checked subtraction. Returns None if overflow occurred.

pub fn abs_diff(self, other: U256) -> U256

Computes the absolute difference between self and other.

pub fn overflowing_mul(self, other: U256) -> (U256, bool)

Multiply with overflow, returning a flag if it does.

pub fn saturating_mul(self, other: U256) -> U256

Multiplication which saturates at the maximum value..

pub fn checked_mul(self, other: U256) -> Option<U256>

Checked multiplication. Returns None if overflow occurred.

pub fn checked_div(self, other: U256) -> Option<U256>

Checked division. Returns None if other == 0.

pub fn checked_rem(self, other: U256) -> Option<U256>

Checked modulus. Returns None if other == 0.

pub fn overflowing_neg(self) -> (U256, bool)

Negation with overflow.

pub fn checked_neg(self) -> Option<U256>

Checked negation. Returns None unless self == 0.

pub fn from_big_endian(slice: &[u8]) -> U256

Converts from big endian representation bytes in memory.

pub fn from_little_endian(slice: &[u8]) -> U256

Converts from little endian representation bytes in memory.

impl U256

pub const fn low_u128(&self) -> u128

Low 2 words (u128)

pub fn as_u128(&self) -> u128

Conversion to u128 with overflow checking

Panics

Panics if the number is larger than 2^128.

impl U256

pub fn full_mul(self, other: U256) -> U512

Multiplies two 256-bit integers to produce full 512-bit integer. Overflow is not possible.