Type Alias ink_primitives::Key

pub type Key = u32;

A key into the smart contract storage.

Note

The storage of an ink! smart contract can be viewed as a key-value store.

In order to manipulate its storage an ink! smart contract is required to indicate the respective cells using this primitive type.

The Key type can be compared to a raw pointer and also allows operations similar to pointer arithmetic.