ink_primitives
Module types

Structs

AccountId

The default environment AccountId type.

EventRecordFoo

Hash

The default environment Hash type.

RuntimeEvent

Enums

DefaultEnvironment

The fundamental types of the default configuration.

DepositLimit

NoChainExtension

Placeholder for chains that have no defined chain extension.

Origin

The type of origins supported by pallet-revive.

Phase

A phase of a block’s execution.

Traits

AccountIdGuard

todo remove A trait to enforce that a type should be an Environment::AccountId.

Clear

The equivalent of Zero for hashes.

CodecAsType

Environment

The environmental types usable by contracts defined with ink!.

FromLittleEndian

Allows to instantiate a type from its little-endian bytes representation.

Type Aliases

Balance

The default balance type.

BlockNumber

The default block number type.

EventRecord

The default event record type.

Gas

The default gas type.

Timestamp

The default timestamp type.